Nuevas bases de datos AD

+++ ACCESO_A_DATOS/TEMA-2/basesDeDatos/hsqldb/src/org/hsqldb/result/ResultLob.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2024, The HSQL Development Group
+/* Copyright (c) 2001-2021, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -47,7 +47,7 @@ import org.hsqldb.rowio.RowOutputInterface;
  * Sub-class of Result for communicating Blob and Clob operations.<p>
  *
  * @author Fred Toussi (fredt@users dot sourceforge.net)
- * @version 2.7.0
+ * @version 2.5.1
  * @since 1.9.0
  */
 public final class ResultLob extends Result {
@@ -104,10 +104,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetBytesRequest(
-            long id,
-            long offset,
-            long length) {
+    public static ResultLob newLobGetBytesRequest(long id, long offset,
+            int length) {
 
         ResultLob result = new ResultLob();
 
@@ -119,10 +117,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetCharsRequest(
-            long id,
-            long offset,
-            long length) {
+    public static ResultLob newLobGetCharsRequest(long id, long offset,
+            int length) {
 
         ResultLob result = new ResultLob();
 
@@ -134,9 +130,7 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobSetBytesRequest(
-            long id,
-            long offset,
+    public static ResultLob newLobSetBytesRequest(long id, long offset,
             byte[] block) {
 
         ResultLob result = new ResultLob();
@@ -150,9 +144,7 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobSetCharsRequest(
-            long id,
-            long offset,
+    public static ResultLob newLobSetCharsRequest(long id, long offset,
             char[] chars) {
 
         ResultLob result = new ResultLob();
@@ -177,9 +169,7 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetBytesResponse(
-            long id,
-            long offset,
+    public static ResultLob newLobGetBytesResponse(long id, long offset,
             byte[] block) {
 
         ResultLob result = new ResultLob();
@@ -193,9 +183,7 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetCharsResponse(
-            long id,
-            long offset,
+    public static ResultLob newLobGetCharsResponse(long id, long offset,
             char[] chars) {
 
         ResultLob result = new ResultLob();
@@ -220,10 +208,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetBytePatternPositionRequest(
-            long id,
-            byte[] pattern,
-            long offset) {
+    public static ResultLob newLobGetBytePatternPositionRequest(long id,
+            byte[] pattern, long offset) {
 
         ResultLob result = new ResultLob();
 
@@ -236,10 +222,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetBytePatternPositionRequest(
-            long id,
-            long otherId,
-            long offset) {
+    public static ResultLob newLobGetBytePatternPositionRequest(long id,
+            long otherId, long offset) {
 
         ResultLob result = new ResultLob();
 
@@ -250,10 +234,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetCharPatternPositionRequest(
-            long id,
-            char[] pattern,
-            long offset) {
+    public static ResultLob newLobGetCharPatternPositionRequest(long id,
+            char[] pattern, long offset) {
 
         ResultLob result = new ResultLob();
 
@@ -266,10 +248,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetCharPatternPositionRequest(
-            long id,
-            long otherId,
-            long offset) {
+    public static ResultLob newLobGetCharPatternPositionRequest(long id,
+            long otherId, long offset) {
 
         ResultLob result = new ResultLob();
 
@@ -281,11 +261,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobCreateBlobRequest(
-            long sessionID,
-            long lobID,
-            InputStream stream,
-            long length) {
+    public static ResultLob newLobCreateBlobRequest(long sessionID,
+            long lobID, InputStream stream, long length) {
 
         ResultLob result = new ResultLob();
 
@@ -297,11 +274,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobCreateClobRequest(
-            long sessionID,
-            long lobID,
-            Reader reader,
-            long length) {
+    public static ResultLob newLobCreateClobRequest(long sessionID,
+            long lobID, Reader reader, long length) {
 
         ResultLob result = new ResultLob();
 
@@ -354,9 +328,7 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLobGetRequest(
-            long id,
-            long offset,
+    public static ResultLob newLobGetRequest(long id, long offset,
             long length) {
 
         ResultLob result = new ResultLob();
@@ -379,10 +351,8 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public static ResultLob newLob(
-            DataInputStream dataInput,
-            boolean readTerminate)
-            throws IOException {
+    public static ResultLob newLob(DataInputStream dataInput,
+                                   boolean readTerminate) throws IOException {
 
         ResultLob result = new ResultLob();
 
@@ -428,7 +398,6 @@ public final class ResultLob extends Result {
                 for (int i = 0; i < result.charBlock.length; i++) {
                     result.charBlock[i] = dataInput.readChar();
                 }
-
                 break;
 
             case LobResultTypes.REQUEST_GET_LENGTH :
@@ -452,7 +421,6 @@ public final class ResultLob extends Result {
                 for (int i = 0; i < result.charBlock.length; i++) {
                     result.charBlock[i] = dataInput.readChar();
                 }
-
                 break;
 
             case LobResultTypes.RESPONSE_SET :
@@ -478,20 +446,16 @@ public final class ResultLob extends Result {
         return result;
     }
 
-    public void write(
-            SessionInterface session,
-            DataOutputStream dataOut,
-            RowOutputInterface rowOut)
-            throws IOException {
+    public void write(SessionInterface session, DataOutputStream dataOut,
+                      RowOutputInterface rowOut) throws IOException {
+
         writeBody(session, dataOut);
         dataOut.writeByte(ResultConstants.NONE);
         dataOut.flush();
     }
 
-    public void writeBody(
-            SessionInterface session,
-            DataOutputStream dataOut)
-            throws IOException {
+    public void writeBody(SessionInterface session,
+                          DataOutputStream dataOut) throws IOException {
 
         switch (subType) {
 
@@ -585,10 +549,8 @@ public final class ResultLob extends Result {
         }
     }
 
-    private void writeCreate(
-            SessionInterface session,
-            DataOutputStream dataOut)
-            throws IOException {
+    private void writeCreate(SessionInterface session,
+                             DataOutputStream dataOut) throws IOException {
 
         dataOut.writeByte(mode);
         dataOut.writeInt(databaseID);
@@ -610,10 +572,9 @@ public final class ResultLob extends Result {
         }
     }
 
-    private void writeCreateByteSegments(
-            SessionInterface session,
-            DataOutputStream dataOut)
-            throws IOException {
+    private void writeCreateByteSegments(SessionInterface session,
+                                         DataOutputStream dataOut)
+                                         throws IOException {
 
         int  bufferLength  = session.getStreamBlockSize();
         long currentOffset = blockOffset;
@@ -624,8 +585,8 @@ public final class ResultLob extends Result {
         dataOut.writeLong(lobID);
         dataOut.writeInt(subType);
 
-        HsqlByteArrayOutputStream byteArrayOS = new HsqlByteArrayOutputStream(
-            bufferLength);
+        HsqlByteArrayOutputStream byteArrayOS =
+            new HsqlByteArrayOutputStream(bufferLength);
 
         byteArrayOS.reset();
         byteArrayOS.write(stream, bufferLength);
@@ -666,10 +627,9 @@ public final class ResultLob extends Result {
         }
     }
 
-    private void writeCreateCharSegments(
-            SessionInterface session,
-            DataOutputStream dataOut)
-            throws IOException {
+    private void writeCreateCharSegments(SessionInterface session,
+                                         DataOutputStream dataOut)
+                                         throws IOException {
 
         int  bufferLength  = session.getStreamBlockSize();
         long currentOffset = blockOffset;
@@ -680,8 +640,8 @@ public final class ResultLob extends Result {
         dataOut.writeLong(lobID);
         dataOut.writeInt(subType);
 
-        HsqlByteArrayOutputStream byteArrayOS = new HsqlByteArrayOutputStream(
-            bufferLength);
+        HsqlByteArrayOutputStream byteArrayOS =
+            new HsqlByteArrayOutputStream(bufferLength);
 
         byteArrayOS.reset();
         byteArrayOS.write(reader, bufferLength / 2);
